Quick test script to understand the API.
The web tool was giving odd results. So I created a quick shell script as
a sanity test. Then, that was giving the same results so I realized that
the mock server occasionally returns results unrelated to the request.
Then I started using the real API and got more sane output so I
expanded the script to help me explore the API a bit more.